Add files via upload
I know a resume in HTML is unnecessary, and often times a template can be plucked from Word or google docs and be done in 5 mins. 

The creation of this resume was to introduce myself to HTML/CSS and uncover its capabilities. Also, to gauge my interest in web creation.
Special thanks to Alvaro Montoro and his article, link here -> https://dev.to/alvaromontoro/developing-an-interactive-resume-with-html-and-css-5007
Drew inspiration from his article, and his resume helped set the layout of how I would like my resume to look like.